story-generator/src/lambdas/fetch_context/requirements.txt
boto3==1.34.69
pinecone-client==3.0.2
langchain==0.1.11
langchain-community==0.0.27
